Add assertion context manager to TestAppContext and convert existing vim tests to use neovim backed test context
