+++ crates/editor/src/editor_tests.rs
@@ -1,20 +1,22 @@
+use std::{cell::RefCell, rc::Rc, time::Instant};
+
+use futures::StreamExt;
+use indoc::indoc;
+use unindent::Unindent;
+
 use super::*;
 use crate::test::{
-    assert_text_with_selections, build_editor, select_ranges, EditorLspTestContext,
-    EditorTestContext,
+    assert_text_with_selections, build_editor, editor_lsp_test_context::EditorLspTestContext,
+    editor_test_context::EditorTestContext, select_ranges,
 };
-use futures::StreamExt;
 use gpui::{
     geometry::rect::RectF,
     platform::{WindowBounds, WindowOptions},
 };
-use indoc::indoc;
 use language::{FakeLspAdapter, LanguageConfig, LanguageRegistry};
 use project::FakeFs;
 use settings::EditorSettings;
-use std::{cell::RefCell, rc::Rc, time::Instant};
 use text::Point;
-use unindent::Unindent;
 use util::{
     assert_set_eq,
     test::{marked_text_ranges, marked_text_ranges_by, sample_text, TextRangeMarker},
