+++ crates/vim/Cargo.toml
@@ -34,6 +34,7 @@ workspace = { path = "../workspace" }
 [dev-dependencies]
 indoc = "1.0.4"
 parking_lot = "0.11.1"
+lazy_static = "1.4"
 
 editor = { path = "../editor", features = ["test-support"] }
 gpui = { path = "../gpui", features = ["test-support"] }
