--- crates/vim/src/test_contexts.rs
@@ -1,9 +1,0 @@
-mod neovim_backed_binding_test_context;
-mod neovim_backed_test_context;
-mod vim_binding_test_context;
-mod vim_test_context;
-
-pub use neovim_backed_binding_test_context::*;
-pub use neovim_backed_test_context::*;
-pub use vim_binding_test_context::*;
-pub use vim_test_context::*;
